Initial Headerfile setting up basic structure of the library.
The intention is that this front end should not need to change much at all,
and the SSAL_internal.h will contain all real computational rountines, this will be further
broken down into sequential, parallel and distributed versions as appropriate. Where appropriate
MKL and/or GSL will be utilised to expliot high perfromance math and RNG functions (however, a pure implementation
will usually be provided to provide the ability to compile without dependencies)

The intention is as this project progresses more routines will be added including options
for spatial and possibly continuous state space problems.